backend/requirements.txt
fastapi
uvicorn
pandas
sqlmodel
python-dotenv
psycopg2-binary
pydantic
